Initial merge work
Make an initial attempt to merge listener and sender.

This is a rough first pass, and there is a lot of room for improvement
and reduction of duplicate effort and code. Channel closing order
seems to work, but could certainly be simplified.

Also removed the 'data' package, since maintaining it was a pain and
it isn't used in any other application anywhere else.